Mock script list / sep
